Create a No-op stream in ClientCallImpl.start() if the Context has been cancelled.
