+++ core/src/main/java/io/grpc/internal/NoopClientStream.java
@@ -1,0 +1,78 @@
+/*
+ * Copyright 2015, Google Inc. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are
+ * met:
+ *
+ *    * Redistributions of source code must retain the above copyright
+ * notice, this list of conditions and the following disclaimer.
+ *    * Redistributions in binary form must reproduce the above
+ * copyright notice, this list of conditions and the following disclaimer
+ * in the documentation and/or other materials provided with the
+ * distribution.
+ *
+ *    * Neither the name of Google Inc. nor the names of its
+ * contributors may be used to endorse or promote products derived from
+ * this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.grpc.internal;
+
+import io.grpc.Compressor;
+import io.grpc.DecompressorRegistry;
+import io.grpc.Status;
+
+import java.io.InputStream;
+
+/**
+ * An implementation of {@link ClientStream} that silently does nothing for the operations.
+ */
+public class NoopClientStream implements ClientStream {
+  public static NoopClientStream INSTANCE = new NoopClientStream();
+
+  @Override
+  public void request(int numMessages) {}
+
+  @Override
+  public void writeMessage(InputStream message) {}
+
+  @Override
+  public void flush() {}
+
+  @Override
+  public boolean isReady() {
+    return false;
+  }
+
+  @Override
+  public void cancel(Status status) {}
+
+  @Override
+  public void halfClose() {}
+
+  @Override
+  public void setCompressor(Compressor c) {
+    // very much a nop
+  }
+
+  @Override
+  public void setDecompressionRegistry(DecompressorRegistry registry) {}
+
+  @Override
+  public void setMessageCompression(boolean enable) {
+    // noop
+  }
+}
